Basic nesting transformer
Provides a walker function and a higher-level Transformer helper
class for extracting the components of a tree in an iterative
(non-recursive) manner.

Went for iteration over recursion because of a preference for
yielding components as they're encountered, which is a lot more
easily done with iterative implementation than recursive.

This has python 3 compatibility issues to be explored.