TokenDict builder func shape change
For the builder callable in TokenDict.from_token to work nicely,
it needs to have the same signature as TokenDict's from_token
in the first place (apart from having a builder option).

So now it does.  This lets us more readily nest other things
in the TokenDict with less fuss.